+++ crud_hotel/src/main/java/com/sh/crud/booking/model/vo/BookingVO.java
@@ -2,20 +2,38 @@ package com.sh.crud.booking.model.vo;
 
 import com.sh.crud.booking.model.entity.Booking;
 import com.sh.crud.member.model.entity.Member;
-import com.sh.crud.pride.model.entity.Pride;
+import com.sh.crud.pride.model.entity.TbPride;
 import com.sh.crud.room.model.entity.Room;
 
-import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
-public class BookingVO extends Booking {
-
+public class BookingVo extends Booking {
+    private List<Room> rooms = new ArrayList<>();
     private Member member;
     private Room room;
-    private Pride pride;
+    private TbPride pride;
 
-    public BookingVO() {
+    public BookingVo() {
     }
 
+    @Override
+    public String toString() {
+        return "BookingVo{" +
+                "rooms=" + rooms +
+                ", member=" + member +
+                ", room=" + room +
+                ", pride=" + pride +
+                '}';
+    }
+
+    public List<Room> getRooms() {
+        return rooms;
+    }
+
+    public void setRooms(List<Room> rooms) {
+        this.rooms = rooms;
+    }
 
     public Member getMember() {
         return member;
@@ -33,27 +51,15 @@ public class BookingVO extends Booking {
         this.room = room;
     }
 
-    public Pride getPride() {
+    public TbPride getPride() {
         return pride;
+
     }
 
-    public void setPride(Pride pride) {
+    public void setPride(TbPride pride) {
         this.pride = pride;
+
     }
 
-    @Override
-    public String toString() {
-        return "BookingVO{" +
-                "bookingNum=" + getBookingNum() +
-                ", bookingMemberId='" + getBookingMemberId() + '\'' +
-                ", bookingRoomNum='" + getBookingRoomNum() + '\'' +
-                ", bookingName='" + getBookingName() + '\'' +
-                ", checkInDate=" + getCheckInDate() +
-                ", checkOutDate=" + getCheckOutDate() +
-                ", bookingTourNum=" + getBookingTourNum() +
-                ", member=" + member +
-                ", room=" + room +
-                ", pride=" + pride +
-                '}';
-    }
+
 }
